requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
numpy==1.26.4
pandas==2.0.2
folium==0.14.0
earthengine-api==0.1.354
joblib==1.2.0
gunicorn==20.1.0
requests==2.31.0
xgboost>=1.7.0
scikit-learn>=1.0.0
google-generativeai